Create test-deployment-pip
