+++ .grabl/automation.yml
@@ -42,9 +42,6 @@ build:
         sudo unlink /usr/bin/python3
         sudo ln -s $(which python3) /usr/bin/python3
         sudo ln -s /usr/share/pyshared/lsb_release.py /opt/pyenv/versions/3.6.10/lib/python3.6/site-packages/lsb_release.py
-        ARTIFACT_USERNAME=$REPO_GRAKN_USERNAME \
-        ARTIFACT_PASSWORD=$REPO_GRAKN_PASSWORD \
-        bazel run @graknlabs_dependencies//distribution/artifact:create-netrc
         bazel build //...
         bazel run @graknlabs_dependencies//tool/checkstyle:test-coverage
         bazel test $(bazel query 'kind(checkstyle_test, //...)') --test_output=errors
@@ -57,9 +54,6 @@ build:
         sudo unlink /usr/bin/python3
         sudo ln -s $(which python3) /usr/bin/python3
         sudo ln -s /usr/share/pyshared/lsb_release.py /opt/pyenv/versions/3.6.10/lib/python3.6/site-packages/lsb_release.py
-        ARTIFACT_USERNAME=$REPO_GRAKN_USERNAME \
-        ARTIFACT_PASSWORD=$REPO_GRAKN_PASSWORD \
-        bazel run @graknlabs_dependencies//distribution/artifact:create-netrc
         bazel test //tests/behaviour/connection/... --test_output=streamed --jobs=1
         bazel test //tests/behaviour/concept/... --test_output=streamed
         bazel test //tests/behaviour/graql/language/match/... --test_output=streamed
@@ -68,46 +62,35 @@ build:
     deploy-pip-snapshot:
       image: graknlabs-ubuntu-20.04
       dependencies: [build, test-behaviour]
-      filter:
-        owner: graknlabs
-        branch: master
+#      filter:
+#        owner: graknlabs
+#        branch: master
       type: foreground
       command: |
         pyenv global 3.6.10
         sudo unlink /usr/bin/python3
         sudo ln -s $(which python3) /usr/bin/python3
         sudo ln -s /usr/share/pyshared/lsb_release.py /opt/pyenv/versions/3.6.10/lib/python3.6/site-packages/lsb_release.py
-        ARTIFACT_USERNAME=$REPO_GRAKN_USERNAME \
-        ARTIFACT_PASSWORD=$REPO_GRAKN_PASSWORD \
-        bazel run @graknlabs_dependencies//distribution/artifact:create-netrc
         export DEPLOY_PIP_USERNAME=$REPO_GRAKN_USERNAME
         export DEPLOY_PIP_PASSWORD=$REPO_GRAKN_PASSWORD
         bazel run --define version=$(git rev-parse HEAD) //:deploy-pip -- snapshot
-      # TODO: fix test-deployment
-#    test-deployment:
-#      image: graknlabs-ubuntu-20.04
-#      dependencies: [deploy-pip-snapshot]
+    test-deployment-pip:
+      image: graknlabs-ubuntu-20.04
+      dependencies: [deploy-pip-snapshot]
 #      filter:
 #        owner: graknlabs
 #        branch: master
-#      type: foreground
-#      command: |
-#        pyenv global 3.6.10
-#        sudo unlink /usr/bin/python3
-#        sudo ln -s $(which python3) /usr/bin/python3
-#        sudo ln -s /usr/share/pyshared/lsb_release.py /opt/pyenv/versions/3.6.10/lib/python3.6/site-packages/lsb_release.py
-#        ARTIFACT_USERNAME=$REPO_GRAKN_USERNAME \
-#        ARTIFACT_PASSWORD=$REPO_GRAKN_PASSWORD \
-#        bazel run @graknlabs_dependencies//distribution/artifact:create-netrc
-#        bazel run //:grakn-extractor -- dist/grakn-core-server-linux
-#        nohup ./dist/grakn-core-server-linux/grakn server start
-#        sleep 60
-#        echo -n "0.0.0-$GRABL_COMMIT" > VERSION
-#        sed -i -e "s/CLIENT_PYTHON_VERSION_MARKER/$(cat VERSION)/g" test/deployment/requirements.txt
-#        cat test/deployment/requirements.txt
-#        pip install --upgrade pip
-#        pip install -r test/deployment/requirements.txt
-#        cd test/deployment/ && python -m unittest test
+      type: foreground
+      command: |
+        pyenv global 3.6.10
+        sudo unlink /usr/bin/python3
+        sudo ln -s $(which python3) /usr/bin/python3
+        sudo ln -s /usr/share/pyshared/lsb_release.py /opt/pyenv/versions/3.6.10/lib/python3.6/site-packages/lsb_release.py
+        bazel run //tests:grakn-extractor -- grakn-core-all-linux
+        ./grakn-core-all-linux/grakn server &
+        pip install --index-url https://repo.grakn.ai/repository/pypi-snapshot/simple grakn-client==$GRABL_COMMIT
+        cd test/deployment/ && python -m unittest test
+
 
 release:
   filter:
